Pull out FakeDAQ to legacy plugins
